+++ packages/datadog_flutter_plugin/lib/src/datadog_sdk_platform_interface.dart
@@ -48,6 +48,7 @@ abstract class DatadogSdkPlatform extends PlatformInterface {
   Future<void> setTrackingConsent(TrackingConsent trackingConsent);
   Future<void> setUserInfo(
       String? id, String? name, String? email, Map<String, Object?> extraInfo);
+  Future<void> addUserExtraInfo(Map<String, Object?> extraInfo);
 
   Future<void> sendTelemetryDebug(String message);
   Future<void> sendTelemetryError(String message, String? stack, String? kind);